Auth0Config type
