+++ packages/auth0/test/auth0.test.ts
@@ -2,22 +2,13 @@ import { describe, it, beforeEach } from '@effection/mocha';
 import expect from 'expect';
 import { createTestServer, Client, Simulation } from './helpers';
 
-import { Auth0Options, createAuth0Simulator} from '../src';
-
-const config: Auth0Options = {
-  scope: 'openid profile email offline_access',
-  port: 4400,
-  audience: "https://thefrontside.auth0.com/api/v1/",
-  tenant: "frontside",
-  clientId: 'IsuLUyWaFczCbAKQrIpVPmyBTFs4g5iq',
-};
-
+import { auth0 } from '../src';
 
 describe('Auth0 simulator', () => {
   let client: Client;
   beforeEach(function*() {
     client = yield createTestServer({
-      simulators: { auth0: createAuth0Simulator(config) }
+      simulators: { auth0 }
     });
   });
