+++ packages/server/src/config/paths.ts
@@ -6,9 +6,7 @@ const certificatesDir = path.join(rootDir, "certs");
 const pemFile = path.join(certificatesDir, "localhost.pem");
 const keyFile = path.join(certificatesDir, "localhost-key.pem");
 
-
 // TODO: need to check certificates exist and if not dispaly meaningful error
-
 
 export const paths = {
   rootDir,
